removed unused interface files and refactored UserInterface, RegisterDataInterface, LoginDataInterface. It's more reacable now
