--- client/src/Interfaces/AuthenticationInterfaces/index.ts
@@ -1,1 +1,0 @@
-export * from './AuthenticationInterfaces';
