+++ client/src/Interfaces/LoginDataInterface/LoginDataInterface.ts
@@ -1,0 +1,4 @@
+export interface LoginDataInterface {
+    email: string;
+    password: string;
+}
