+++ client/src/Interfaces/LoginDataInterface/index.ts
@@ -1,0 +1,1 @@
+export * from './LoginDataInterface';
