+++ client/src/Interfaces/RegisterDataInterface/RegisterDataInterface.ts
@@ -1,0 +1,9 @@
+export interface RegisterDataInterface {
+    firstName?: string;
+    lastName?: string;
+    email?: string;
+    address?: string;
+    phone?: string;
+    password?: string;
+    confirmPassword?: string;
+}
