+++ client/src/Interfaces/RegisterDataInterface/index.ts
@@ -1,0 +1,1 @@
+export * from "./RegisterDataInterface";
