+++ client/src/Interfaces/UserInterface/UserInterface.ts
@@ -1,0 +1,9 @@
+export interface UserInterface {
+    firstName: string;
+    lastName: string;
+    email: string;
+    address: string;
+    phone: string;
+    token?: string;
+    isAuthenticated: boolean;
+}
