+++ client/src/Interfaces/UserInterface/index.ts
@@ -1,0 +1,1 @@
+export * from './UserInterface';
